attribute/M3DFEL/requirements.txt
torch
timm
einops
sklearn
seaborn
matplotlib
numpy
torchvision
pillow